requirements.txt
albumentations==0.5.2
fvcore==0.1.5.post20220512
numpy==1.21.2
opencv_python==4.5.3.56
ptflops==0.6.9
PyYAML==6.0
scikit_image==0.18.3
scipy==1.7.1
skimage==0.0
timm==0.4.12
torch==1.9.1+cu111
torchvision==0.10.1+cu111
